subsurface scattering in grass and trees!

+++ Minecraft Clone/Shaders/PackedBlock.vert
@@ -18,8 +18,8 @@ out vec2 texCoord;
 out vec3 vNormal;
 out vec4 worldPos;
 out float vertexBrightness;
-out int isWater;
-out int isFoliage;
+out float isWater;
+out float isFoliage;
 
 vec3 DecodePos(uint p) {
     float x = ((p >> 0) & 0x7Fu);
@@ -51,7 +51,6 @@ void main()
 
     isWater   = int(wigX);
     isFoliage = int(wigY);
-
 
     vec4 position = vec4(DecodePos(inPosNorBright), 1.0);
     worldPos = model * position;
